backend/members/requirements.txt
pymongo
strawberry-graphql[cli]
strawberry-graphql[fastapi]
strawberry-graphql[artwork]
strawberry-graphql[strawberry]
fastapi
uvicorn
pydantic